add centralized/decentralized option for main server
